nutrivision/backend/nutrivision-api/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
boto3==1.28.44
pydantic==1.10.13
mangum==0.17.0
requests==2.31.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0